# openapi.yaml
{
  "components": {
    "schemas": {
      "AssignRoleRequest": {
        "properties": {
          "role": {
            "enum": [
              "user",
              "admin",
              "content_manager",
              "auditor",
              "support"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "AssignRoleRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "RoleAssignmentsResponse": {
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "roles"
        ],
        "title": "RoleAssignmentsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Auth Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/users/{user_id}/block": {
      "post": {
        "operationId": "block_user_v1_admin_users__user_id__block_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Block User",
        "tags": [
          "admin",
          "admin"
        ]
      }
    },
    "/v1/admin/users/{user_id}/roles": {
      "get": {
        "operationId": "list_roles_v1_admin_users__user_id__roles_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "List Roles",
        "tags": [
          "admin",
          "admin"
        ]
      },
      "post": {
        "operationId": "assign_role_v1_admin_users__user_id__roles_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Assign Role",
        "tags": [
          "admin",
          "admin"
        ]
      }
    },
    "/v1/admin/users/{user_id}/sessions": {
      "get": {
        "operationId": "list_user_sessions_v1_admin_users__user_id__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  },
                  "title": "Response List User Sessions V1 Admin Users  User Id  Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "List User Sessions",
        "tags": [
          "admin",
          "admin"
        ]
      }
    },
    "/v1/admin/users/{user_id}/sessions/{token_id}/revoke": {
      "post": {
        "operationId": "revoke_user_session_v1_admin_users__user_id__sessions__token_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Revoke User Session",
        "tags": [
          "admin",
          "admin"
        ]
      }
    },
    "/v1/admin/users/{user_id}/unblock": {
      "post": {
        "operationId": "unblock_user_v1_admin_users__user_id__unblock_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Unblock User",
        "tags": [
          "admin",
          "admin"
        ]
      }
    },
...
  }
}